second commit

+++ back/base/models.py
@@ -66,7 +66,7 @@ class Order(models.Model):
     id = models.AutoField(primary_key=True,editable=False)
     user = models.ForeignKey(User,on_delete=models.SET_NULL,null=True)
     createdTime = models.DateField(auto_now_add=True)
-    total = models.IntegerField(null=True)
+    total = models.DecimalField(max_digits=5,decimal_places=2,null=True)
 
 class OrderSerializer(serializers.ModelSerializer):
     class Meta:
@@ -79,7 +79,7 @@ class Order_Details(models.Model):
     user = models.ForeignKey(User,on_delete=models.SET_NULL,null=True)
     product = models.ForeignKey(Product,on_delete=models.SET_NULL,null=True)
     amount = models.IntegerField(null=True)
-    total = models.IntegerField(null=True)
+    total = models.DecimalField(max_digits=5,decimal_places=2,null=True)
 
 class OrderDetSerializer(serializers.ModelSerializer):
     class Meta:
